Do some refactoring on v1 handlers under /api/v1/w/[wId]/assistant
