+++ front/lib/api/assistant/pubsub.ts
@@ -189,6 +189,20 @@ export async function interruptAgentLoop(
   });
 }
 
+export type MessageStreamEvent = {
+  eventId: string;
+  data: (
+    | AgentErrorEvent
+    | AgentActionRunningEvents
+    | AgentActionSuccessEvent
+    | AgentGenerationCancelledEvent
+    | AgentToolCallStartedEvent
+    | GenerationTokensEvent
+  ) & {
+    step: number;
+  };
+};
+
 export async function* getMessagesEvents(
   auth: Authenticator,
   {
@@ -196,22 +210,7 @@ export async function* getMessagesEvents(
     lastEventId,
     signal,
   }: { messageId: string; lastEventId: string | null; signal: AbortSignal }
-): AsyncGenerator<
-  {
-    eventId: string;
-    data: (
-      | AgentErrorEvent
-      | AgentActionRunningEvents
-      | AgentActionSuccessEvent
-      | AgentGenerationCancelledEvent
-      | AgentToolCallStartedEvent
-      | GenerationTokensEvent
-    ) & {
-      step: number;
-    };
-  },
-  void
-> {
+): AsyncGenerator<MessageStreamEvent, void> {
   const pubsubChannel = getMessageChannelId(messageId);
 
   const start = Date.now();
